+++ lib/sequent/generator.rb
@@ -1,0 +1,2 @@
+require_relative './generator/project'
+require_relative './generator/aggregate'