+++ lib/sequent/generator/template_aggregate/template_aggregate/commands.rb
@@ -1,0 +1,2 @@
+class AddTemplateAggregate < Sequent::Command
+end