+++ lib/sequent/generator/template_aggregate/template_aggregate/events.rb
@@ -1,0 +1,2 @@
+class TemplateAggregateAdded < Sequent::Event
+end