+++ lib/sequent/generator/template_aggregate/template_aggregate/template_aggregate.rb
@@ -1,0 +1,9 @@
+class TemplateAggregate < Sequent::AggregateRoot
+  def initialize(command)
+    super(command.aggregate_id)
+    apply TemplateAggregateAdded
+  end
+
+  on TemplateAggregateAdded do
+  end
+end