+++ lib/sequent/generator/template_project/lib/post.rb
@@ -1,0 +1,4 @@
+require_relative 'post/commands'
+require_relative 'post/events'
+require_relative 'post/post'
+require_relative 'post/post_command_handler'