+++ lib/sequent/generator/template_project/my_app.rb
@@ -1,0 +1,11 @@
+require 'sequent'
+require 'sequent/support'
+require 'erb'
+require_relative 'lib/post'
+require_relative 'app/projectors/post_projector'
+
+require_relative 'config/initializers/sequent'
+
+module MyApp
+
+end