+++ lib/sequent/sequent.rb
@@ -1,23 +1,37 @@
-require_relative 'core/core'
-require_relative 'util/util'
-require_relative 'migrations/migrations'
 require_relative 'configuration'
-
-require 'logger'
+require_relative 'core/event'
+require_relative 'core/command'
+require_relative 'core/base_command_handler'
+require_relative 'core/aggregate_root'
+require_relative 'core/projector'
+require_relative 'core/workflow'
+require_relative 'generator'
 
 module Sequent
   def self.new_uuid
     Sequent.configuration.uuid_generator.uuid
   end
 
-  def self.logger
-    @logger ||= Logger.new(STDOUT).tap {|l| l.level = Logger::INFO }
-  end
-
-  def self.logger=(logger)
-    @logger = logger
-  end
-
+  #
+  # Setup Sequent.
+  #
+  # Setup is typically called in an +initializer+ or setup phase of your application.
+  # A minimal setup could look like this:
+  #
+  #   Sequent.configure do |config|
+  #     config.event_handlers = [
+  #       MyProjector.new,
+  #       AnotherProjector.new,
+  #       MyWorkflow.new,
+  #     ]
+  #
+  #     config.command_handlers = [
+  #       MyCommandHandler.new,
+  #     ]
+  #
+  #   end
+  #
+  #
   def self.configure
     yield Configuration.instance
   end
@@ -30,4 +44,30 @@ module Sequent
   def self.command_service
     configuration.command_service
   end
+
+  def self.new_version
+    migration_class.version
+  end
+
+  def self.migration_class
+    Class.const_get(configuration.migrations_class_name)
+  end
+
+  # Short hand for Sequent.configuration.logger
+  def self.logger
+    configuration.logger
+  end
+
+  # Short hand for Sequent.configuration.aggregate_repository
+  def self.aggregate_repository
+    configuration.aggregate_repository
+  end
+
+  # Shortcut classes for easy usage
+  Event = Sequent::Core::Event
+  Command = Sequent::Core::Command
+  CommandHandler = Sequent::Core::BaseCommandHandler
+  AggregateRoot = Sequent::Core::AggregateRoot
+  Projector = Sequent::Core::Projector
+  Workflow = Sequent::Core::Workflow
 end